Refactor RazorEditService.CSharpMember creation
