+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/DocumentMapping/RazorEditService_Members.cs
@@ -10,7 +10,6 @@ using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Language.Extensions;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
 using Microsoft.AspNetCore.Razor.PooledObjects;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Razor.Formatting;
 using Microsoft.CodeAnalysis.Razor.Protocol;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
@@ -197,66 +196,13 @@ internal partial class RazorEditService
         using var members = new PooledArrayBuilder<CSharpMember>();
         foreach (var member in classDecl.Members)
         {
-            if (TryCreateMember(member, sourceText) is { } csharpMember)
+            if (CSharpMember.TryCreate(member, sourceText) is { } csharpMember)
             {
                 members.Add(csharpMember);
             }
         }
 
         return members.ToImmutableAndClear();
-    }
-
-    private static CSharpMember? TryCreateMember(MemberDeclarationSyntax member, SourceText sourceText)
-        => member switch
-        {
-            MethodDeclarationSyntax method => new(method, GetComparisonSpan(method), sourceText),
-            PropertyDeclarationSyntax property => new(property, GetComparisonSpan(property), sourceText),
-            FieldDeclarationSyntax field => new(field, GetComparisonSpan(field), sourceText),
-            _ => null,
-        };
-
-    private static TextSpan GetComparisonSpan(MethodDeclarationSyntax method)
-    {
-        // Since we only want to know about additions, we need to ignore any body changes, so we end our comparison span
-        // before the body, or expression body, starts. This prevents changes inside method bodies that are entirely unmapped
-        // causing us to add that method. Since an existing unmapped method can only be present if the Razor compiler emitted
-        // it, we never want those in the Razor file.
-        // Strictly speaking this is comparing more than necessary - since a C# method can't be overloaded by return type for
-        // example, having that as part of the comparison is redundant. Same for visibility modifiers, which would seem to show
-        // a bug in this logic: If Roslyn changes a method from public to private via a code action, that would appear to this
-        // logic as an addition. In reality though, such a change would have to be in a mappable region to be a valid code action,
-        // so the edits will have been processed already, and not seen by this code. For a method to go from public to private
-        // in an unmappable region means Roslyn is changing one of the Razor compiler generated methods, which the user can
-        // never see or interact with.
-        // If the user has an incomplete method, then we are safe to just use the end of the method node.
-        if (((SyntaxNode?)method.Body ?? method.ExpressionBody)?.SpanStart is not { } spanEnd)
-        {
-            spanEnd = method.Span.End;
-        }
-
-        return TextSpan.FromBounds(method.SpanStart, spanEnd);
-    }
-
-    private static TextSpan GetComparisonSpan(PropertyDeclarationSyntax property)
-    {
-        // Properties can't be overloaded, so the name alone is enough to tell whether a generated property
-        // already exists. Keeping the comparison this narrow avoids treating accessor, initializer, or
-        // modifier changes as additions.
-        return property.Identifier.Span;
-    }
-
-    private static TextSpan GetComparisonSpan(FieldDeclarationSyntax field)
-    {
-        // Fields can't be overloaded either, so the declared variable name(s) are enough to identify an
-        // existing generated field. Comparing only that span avoids treating modifier or initializer changes
-        // as additions.
-        var variables = field.Declaration.Variables;
-        if (variables.Count == 0)
-        {
-            return field.Declaration.Span;
-        }
-
-        return TextSpan.FromBounds(variables[0].Identifier.SpanStart, variables[^1].Identifier.Span.End);
     }
 
     private static bool IsLineEmpty(TextLine textLine)
